requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
isort==5.9.3
mccabe==0.6.1
pep8-naming==0.12.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
testfixtures==6.18.2
